requirements.txt
jinja2
pdfkit
scikit-learn
numpy
pandas
python-docx
pdfplumber
spacy